add additional TableProperties; implement TableAutoformatLookSpecifier
git-svn-id: https://svn.apache.org/repos/asf/poi/trunk@1142827 13f79535-47bb-0310-9956-ffa450edef68

+++ src/scratchpad/src/org/apache/poi/hwpf/usermodel/TableRow.java
@@ -82,7 +82,7 @@ public final class TableRow
 
   public void setRowJustification(int jc)
   {
-    _tprops.setJc(jc);
+    _tprops.setJc((short) jc);
     _papx.updateSprm(SPRM_TJC, (short)jc);
   }
